+++ mobile/features/bookmarks/types/index.ts
@@ -1,4 +1,4 @@
-import { NewsItem } from "./newsItem";
+import { NewsItem } from "@/features/news/types";
 
 export interface BookmarkState {
     [key: string]: NewsItem;
